Interpolatin fixed, capacitor currents still to go

+++ main_project/output/nonLinearAnalysis.hpp
@@ -5,7 +5,7 @@
 #include <circuit/circuit.hpp>
 
 void nonLinearSetup(Circuit& c, bool isDc = false);
-string runNonLinearTransience(Circuit& c, float t, VectorXd& interpolX1, VectorXd& interpolX2, double& interpolT1, double& interpolT2,double& printTime,vector<float>& interpolI1, vector<float>& interpolI2);
+string runNonLinearTransience(Circuit& c, double t, VectorXd& interpolX1, VectorXd& interpolX2, double& interpolT1, double& interpolT2,double& printTime,vector<float>& interpolI1, vector<float>& interpolI2);
 bool matrixDiffBellowThreshold(VectorXd& m1, VectorXd& m2, float d);
 void initializeDcBias(Circuit &c, int maxIterationsPerSourceStep = 5, float minimumStep = 1e-4, float threshold = 0.001);
 bool hasNan(VectorXd& m);
